Smoothing rough edges on kafka logging
