+++ util/logging/logging.go
@@ -135,14 +135,6 @@ func newls(name string, level Level, bundle *dumpBundle) *LogSet {
 
 // Configure allows an existing LogSet to change its settings.
 func (ls *LogSet) Configure(cfg Config) error {
-	ls.logrus.SetLevel(cfg.getLogrusLevel())
-
-	if cfg.Basic.DisableConsole {
-		ls.dumpBundle.err = ioutil.Discard
-	} else {
-		ls.dumpBundle.err = ls.dumpBundle.defaultErr
-	}
-
 	err := ls.configureKafka(cfg)
 	if err != nil {
 		return err
@@ -153,48 +145,16 @@ func (ls *LogSet) Configure(cfg Config) error {
 		return err
 	}
 
+	ls.logrus.SetLevel(cfg.getLogrusLevel())
+
+	if cfg.Basic.DisableConsole {
+		ls.dumpBundle.err = ioutil.Discard
+	} else {
+		ls.dumpBundle.err = ls.dumpBundle.defaultErr
+	}
+
 	ls.liveConfig = &cfg
 	return nil
-}
-
-type kafkaConfigurationMessage struct {
-	CallerInfo
-	hook    *kafkalogrus.KafkaLogrusHook
-	brokers []string
-	topic   string
-}
-
-func reportKafkaConfig(hook *kafkalogrus.KafkaLogrusHook, cfg Config, ls LogSink) {
-	msg := kafkaConfigurationMessage{
-		CallerInfo: GetCallerInfo(),
-		hook:       hook,
-		brokers:    cfg.getBrokers(),
-		topic:      cfg.Kafka.Topic,
-	}
-	Deliver(msg, ls)
-}
-
-func (kcm kafkaConfigurationMessage) DefaultLevel() Level {
-	return InformationLevel
-}
-
-func (kcm kafkaConfigurationMessage) Message() string {
-	if kcm.hook == nil {
-		return "Not connecting to Kafka."
-	}
-	return "Connecting to Kafka"
-}
-
-func (kcm kafkaConfigurationMessage) EachField(f FieldReportFn) {
-	f("@loglov3-otl", "sous-kafka-config")
-	kcm.CallerInfo.EachField(f)
-	if kcm.hook == nil {
-		return
-	}
-	f("logging-topic", kcm.topic)
-	f("brokers", kcm.brokers)
-	f("logger-id", kcm.hook.Id())
-	f("levels", kcm.hook.Levels())
 }
 
 func (ls LogSet) configureKafka(cfg Config) error {
@@ -228,45 +188,11 @@ func (ls LogSet) configureKafka(cfg Config) error {
 	return nil
 }
 
-type graphiteConfigMessage struct {
-	CallerInfo
-	cfg *graphite.Config
-}
-
-func reportGraphiteConfig(cfg *graphite.Config, ls LogSink) {
-	msg := graphiteConfigMessage{
-		CallerInfo: GetCallerInfo(),
-		cfg:        cfg,
-	}
-	Deliver(msg, ls)
-}
-
-func (gcm graphiteConfigMessage) DefaultLevel() Level {
-	return InformationLevel
-}
-
-func (gcm graphiteConfigMessage) Message() string {
-	if gcm.cfg == nil {
-		return "Not connecting to Graphite server"
-	}
-	return "Connecting to Graphite server"
-}
-
-func (gcm graphiteConfigMessage) EachField(f FieldReportFn) {
-	f("@loglov3-otl", "sous-graphite-config")
-	gcm.CallerInfo.EachField(f)
-	if gcm.cfg == nil {
-		return
-	}
-	f("server-addr", gcm.cfg.Addr)
-	f("flush-interval", gcm.cfg.FlushInterval)
-}
-
 func (ls LogSet) configureGraphite(cfg Config) error {
 	var gCfg *graphite.Config
 
 	if cfg.useGraphite() {
-		addr, err := net.ResolveTCPAddr("tcp", cfg.Graphite.Server)
+		addr, err := net.ResolveTCPAddr("tcp", cfg.getGraphiteServer())
 		if err != nil {
 			return err
 		}
